requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
pytest==7.4.3
pytest-cov==4.1.0
